src/pyproject.toml
[build-system]
requires = [ "setuptools>=42", "wheel",]
build-backend = "setuptools.build_meta"

[project]
name = "gcp-logger"
version = "0.1.0"
description = "A flexible Python package for logging with Google Cloud integration and multi-environment support."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [ "colorama==0.4.6", "google-cloud-logging==3.11.2", "gcloud-aio-storage==9.3.0",]
[[project.authors]]
name = "Caio Pizzol"
email = "caio@harbourshare.com"

[project.license]
text = "MIT"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--maxfail=2 -ra"
testpaths = [ "tests",]
